EditProductForm complete

+++ src/main/java/com/example/inventorymanagementsystem/controllers/MainFormController.java
@@ -3,10 +3,11 @@ package com.example.inventorymanagementsystem.controllers;
 import com.example.inventorymanagementsystem.data.Part;
 import com.example.inventorymanagementsystem.data.Product;
 import javafx.application.Platform;
+import javafx.beans.InvalidationListener;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Label;
 import javafx.scene.control.SelectionModel;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
@@ -25,10 +26,21 @@ public class MainFormController extends BaseController {
     @FXML
     private TextField searchProductsTextField;
 
+    @FXML
+    private Label searchPartsByIdErrorLabel;
+
+    @FXML
+    private Label searchProductsByIdErrorLabel;
+
     @Override
     public void setupUI() {
         setupPartsTableView();
         setupProductsTableView();
+    }
+
+    @Override
+    public void tearDown() {
+
     }
 
     private void setupPartsTableView() {
@@ -49,6 +61,9 @@ public class MainFormController extends BaseController {
         stockColumn.setCellValueFactory(new PropertyValueFactory<>("stock"));
 
         partsTableView.setItems(getInventory().getAllParts());
+        getInventory().getAllParts().addListener((InvalidationListener) observable -> {
+            refreshParts();
+        });
         partsTableView.getColumns().addAll(idColumn, nameColumn, stockColumn, priceColumn);
         partsTableView.setEditable(false);
     }
@@ -71,16 +86,22 @@ public class MainFormController extends BaseController {
         priceColumn.setCellValueFactory(new PropertyValueFactory<>("price"));
 
         productsTableView.setItems(getInventory().getAllProducts());
+        getInventory().getAllProducts().addListener((InvalidationListener) observable -> {
+            refreshProducts();
+        });
         productsTableView.getColumns().addAll(idColumn, nameColumn, stockColumn, priceColumn);
         productsTableView.setEditable(false);
     }
 
-    public void searchParts() {
+    private void refreshParts() {
         String query = searchPartsTextField.getText();
+        searchPartsByIdErrorLabel.setVisible(false);
         // Inventory really should handle this logic of whether to search via id or not, but I'm adhering to the UML diagrams
         try {
             int partId = Integer.parseInt(query);
-            partsTableView.setItems(FXCollections.observableList(List.of(getInventory().lookupPart(partId))));
+            List<Part> queryResults = List.of(getInventory().lookupPart(partId));
+            partsTableView.setItems(FXCollections.observableList(queryResults));
+            searchPartsByIdErrorLabel.setVisible(query.isEmpty());
         } catch (NumberFormatException exception) {
             ObservableList<Part> parts = getInventory().lookupPart(query);
             partsTableView.setItems(parts);
@@ -88,40 +109,51 @@ public class MainFormController extends BaseController {
             partsTableView.setItems(getInventory().getAllParts());
         }
     }
+
+    public void searchParts() {
+        refreshParts();
+    }
     public void addPart() {
         getScreenNavigator().switchToEditPartForm(-1,null);
     }
-    public void modifyPart(ActionEvent e) {
+    public void modifyPart() {
         SelectionModel<Part> selectionModel = partsTableView.getSelectionModel();
         Part selectedPart = selectionModel.getSelectedItem();
         int selectedIndex = selectionModel.getSelectedIndex();
         if (selectedPart != null)
             getScreenNavigator().switchToEditPartForm(selectedIndex, selectedPart);
     }
-    public void deletePart(ActionEvent e) {
-        getDialogManager().showDeleteConfirmationDialog(() -> {
-            Part selectedPart = partsTableView.getSelectionModel().getSelectedItem();
-            if (selectedPart != null)
+    public void deletePart() {
+        Part selectedPart = partsTableView.getSelectionModel().getSelectedItem();
+        if (selectedPart != null) {
+            getDialogManager().showDeleteConfirmationDialog(() -> {
                 getInventory().deletePart(selectedPart);
-            // Shouldn't be doing this manually
-            searchParts();
-        }, null);
+            }, null);
+        }
     }
     public void exit() {
         getDialogManager().showExitConfirmationDialog(Platform::exit, null);
     }
-    public void searchProducts() {
+
+    private void refreshProducts() {
         String query = searchProductsTextField.getText();
+        searchProductsByIdErrorLabel.setVisible(false);
         // Inventory really should handle this logic of whether to search via id or not, but I'm adhering to the UML diagrams
         try {
             int productId = Integer.parseInt(query);
-            productsTableView.setItems(FXCollections.observableList(List.of(getInventory().lookupProduct(productId))));
+            List<Product> queryResults = List.of(getInventory().lookupProduct(productId));
+            productsTableView.setItems(FXCollections.observableList(queryResults));
+            searchProductsByIdErrorLabel.setVisible(queryResults.isEmpty());
         } catch (NumberFormatException exception) {
             ObservableList<Product> products = getInventory().lookupProduct(query);
             productsTableView.setItems(products);
         } catch (Exception exception) {
             productsTableView.setItems(getInventory().getAllProducts());
         }
+    }
+
+    public void searchProducts() {
+        refreshProducts();
     }
     public void addProduct() {
         getScreenNavigator().switchToEditProductForm(-1,null);
@@ -134,12 +166,14 @@ public class MainFormController extends BaseController {
             getScreenNavigator().switchToEditProductForm(selectedIndex, selectedProduct);
     }
     public void deleteProduct() {
-        getDialogManager().showDeleteConfirmationDialog(() -> {
-            Product selectedProduct = productsTableView.getSelectionModel().getSelectedItem();
-            if (selectedProduct != null)
-                getInventory().deleteProduct(selectedProduct);
-            // A manual refresh shouldn't be necessary
-            searchProducts();
-        }, null);
+        Product selectedProduct = productsTableView.getSelectionModel().getSelectedItem();
+        if (selectedProduct != null) {
+            getDialogManager().showDeleteConfirmationDialog(() -> {
+                if (getInventory().lookupProduct(selectedProduct.getId()).getAllAssociatedParts().isEmpty())
+                    getInventory().deleteProduct(selectedProduct);
+                else
+                    getDialogManager().showAlertDialog("Cannot delete, please remove associated parts before deleting this product.");
+            }, null);
+        }
     }
 }
